ml_project/requirements.txt
pandas~=1.1.5
scikit-learn~=0.24.2
dataclasses~=0.8
pyyaml~=5.4.1
numpy~=1.19.5
click~=7.1.2
setuptools~=52.0.0
faker~=8.1.2
pandas_profiling~=2.12.0
jupyterlab~=3.0.14
ipywidgets~=7.6.3